Add cmake support, fix problems with building with visual studio, move include to proper folder structure.
