+++ groundupdb/include/groundupdb/database.h
@@ -1,6 +1,7 @@
 #ifndef DATABASE_H
 #define DATABASE_H
 
+#include <memory>
 #include <string>
 
 namespace groundupdb {
